fix: Updated contracts to better align with nitro deploy tooling and fixed replay bugs
