+++ deploy/deploy.go
@@ -10,6 +10,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/log"
 	"github.com/offchainlabs/nitro/cmd/chaininfo"
 	"github.com/offchainlabs/nitro/solgen/go/bridgegen"
 	"github.com/offchainlabs/nitro/solgen/go/challengegen"
@@ -41,27 +42,12 @@ func deployBridgeCreator(ctx context.Context, l1Reader *headerreader.HeaderReade
 		return common.Address{}, fmt.Errorf("bridge deploy error: %w", err)
 	}
 
-	dummyRollupManager, tx, _, err := bridgegen.DeployEigenDADummyManager(auth, client)
-	err = andTxSucceeded(ctx, l1Reader, tx, err)
-	if err != nil {
-		return common.Address{}, fmt.Errorf("dummy manager deploy error: %w", err)
-	}
-
-	println("Dummy manager deployed at ", dummyRollupManager.String())
-
-	dummySvcManager, tx, _, err := bridgegen.DeployDummyServiceManager(auth, client)
-	err = andTxSucceeded(ctx, l1Reader, tx, err)
-	if err != nil {
-		return common.Address{}, fmt.Errorf("dummy svc manager deploy error: %w", err)
-	}
-
-	println("Dummy service manager deployed at ", dummySvcManager.String())
 	reader4844, tx, _, err := yulgen.DeployReader4844(auth, client)
 	err = andTxSucceeded(ctx, l1Reader, tx, err)
 	if err != nil {
 		return common.Address{}, fmt.Errorf("blob basefee reader deploy error: %w", err)
 	}
-	seqInboxTemplate, tx, _, err := bridgegen.DeploySequencerInbox(auth, client, maxDataSize, reader4844, dummySvcManager, dummyRollupManager, isUsingFeeToken)
+	seqInboxTemplate, tx, _, err := bridgegen.DeploySequencerInbox(auth, client, maxDataSize, reader4844, isUsingFeeToken)
 	err = andTxSucceeded(ctx, l1Reader, tx, err)
 	if err != nil {
 		return common.Address{}, fmt.Errorf("sequencer inbox deploy error: %w", err)
@@ -251,11 +237,37 @@ func deployRollupCreator(ctx context.Context, l1Reader *headerreader.HeaderReade
 	return rollupCreator, rollupCreatorAddress, validatorUtils, validatorWalletCreator, nil
 }
 
-func DeployOnL1(ctx context.Context, parentChainReader *headerreader.HeaderReader, deployAuth *bind.TransactOpts, batchPosters []common.Address, batchPosterManager common.Address, authorizeValidators uint64, config rollupgen.Config, nativeToken common.Address, maxDataSize *big.Int, isUsingFeeToken bool) (*chaininfo.RollupAddresses, error) {
+func DeployOnL1(ctx context.Context, parentChainReader *headerreader.HeaderReader, deployAuth *bind.TransactOpts, batchPosters []common.Address, batchPosterManager common.Address, authorizeValidators uint64, config rollupgen.Config, nativeToken common.Address, maxDataSize *big.Int, isUsingFeeToken bool, eigenDASvcManager common.Address, eigenDARollupManager common.Address) (*chaininfo.RollupAddresses, error) {
 	if config.WasmModuleRoot == (common.Hash{}) {
 		return nil, errors.New("no machine specified")
 	}
 
+	if eigenDARollupManager == (common.Address{0x0}) {
+		log.Warn("No EigenDA Rollup Manager contract address specified, deploying dummy rollup manager instead")
+
+		dummyRollupManager, tx, _, err := bridgegen.DeployEigenDADummyManager(deployAuth, parentChainReader.Client())
+		err = andTxSucceeded(ctx, parentChainReader, tx, err)
+		if err != nil {
+			return nil, fmt.Errorf("dummy manager deploy error: %w", err)
+		}
+
+		log.Info("Dummy manager deployed at ", dummyRollupManager.String())
+		eigenDARollupManager = dummyRollupManager
+	}
+
+	if eigenDASvcManager == (common.Address{0x0}) {
+		log.Warn("No EigenDA Service Manager contract address specified, deploying dummy service manager instead")
+
+		dummySvcManager, tx, _, err := bridgegen.DeployDummyServiceManager(deployAuth, parentChainReader.Client())
+		err = andTxSucceeded(ctx, parentChainReader, tx, err)
+		if err != nil {
+			return nil, fmt.Errorf("dummy svc manager deploy error: %w", err)
+		}
+
+		log.Info("Dummy service manager deployed at ", dummySvcManager.String())
+		eigenDASvcManager = dummySvcManager
+
+	}
 	rollupCreator, _, validatorUtils, validatorWalletCreator, err := deployRollupCreator(ctx, parentChainReader, deployAuth, maxDataSize, isUsingFeeToken)
 	if err != nil {
 		return nil, fmt.Errorf("error deploying rollup creator: %w", err)
@@ -275,6 +287,8 @@ func DeployOnL1(ctx context.Context, parentChainReader *headerreader.HeaderReade
 		MaxFeePerGasForRetryables: big.NewInt(0), // needed when utility factories are deployed
 		BatchPosters:              batchPosters,
 		BatchPosterManager:        batchPosterManager,
+		EigenDAServiceManager:     eigenDASvcManager,
+		EigenDARollupManager:      eigenDARollupManager,
 	}
 
 	tx, err := rollupCreator.CreateRollup(
