+++ eigenda/eigenda_test.go
@@ -1,0 +1,34 @@
+package eigenda
+
+import (
+	"encoding/hex"
+	"testing"
+)
+
+func TestParseSequencerMsg(t *testing.T) {
+	calldataString := "43a85289000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000000000000000000000000000340000000000000000000000000000000000000000000000000000000000000000300000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000000000000000000000004500000000000000000000000000000000000000000000000000000000000001a400000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000024000000000000000000000000000000000000000000000000000000000000000603110cdec4dda50a1465d571f980c07db659331ffc7b376d5eef7fe298ded2f3f00000000000000000000000000000000000000000000000000000000000000043110cdec4dda50a1465d571f980c07db659331ffc7b376d5eef7fe298ded2f3f000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000002010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000026090000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000a54baeff59fa897b3360d0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000020100000000000000000000000000000000000000000000000000000000000000007e2db2683cd5ec31b62b50b9a685140076b483f1f85b931f493480cbfd9eda10a964fcc86dbace6cedd749b878523e8bdc8ad1c04104cdbf1482d79e3367b90000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000001d000000000000000000000000000000000000000000000000000000000000001e000000000000000000000000000000000000000000000000000000000000012c"
+
+	calldata, err := hex.DecodeString(calldataString)
+	if err != nil {
+		t.Fatalf("Failed to decode calldata: %v", err)
+	}
+
+	expected := &EigenDABlobInfo{
+		// BatchHeader content for hashing
+		BlobVerificationProof: BlobVerificationProof{
+			BatchID: 69,
+		},
+	}
+
+	// Call the function with the mock calldata
+	result, err := ParseSequencerMsg(calldata)
+	if err != nil {
+		t.Fatalf("ParseSequencerMsg returned an error: %v", err)
+	}
+
+	// TODO: Extend the test to cover all fields
+	if result.BlobVerificationProof.BatchID != expected.BlobVerificationProof.BatchID {
+		t.Errorf("BlobIndex was incorrect, got: %v, want: %v", result.BlobVerificationProof.BatchID, expected.BlobVerificationProof.BatchID)
+	}
+
+}
