+++ staker/stateless_block_validator.go
@@ -343,13 +343,12 @@ func (v *StatelessBlockValidator) ValidationEntryRecord(ctx context.Context, e *
 			if v.eigenDAService == nil {
 				log.Warn("EigenDA not configured, but sequencer message found with EigenDA header")
 			} else {
-
-				println("RecoverPayloadFromEigenDABatch .... recovering payload for EigenDA batch")
 				// we use the polynomial domain here because this is what we use in the fraud proof pipeline
-				_, err := eigenda.RecoverPayloadFromEigenDABatch(ctx, batch.Data[41:], v.eigenDAService, e.Preimages, "polynomial")
+				_, err := eigenda.RecoverPayloadFromEigenDABatch(ctx, batch.Data[41:], v.eigenDAService, e.Preimages, "binary")
 				if err != nil {
 					return err
 				}
+				log.Info("Recovered polynomial from EigenDA batch", "batch", batch.Number)
 			}
 		}
 	}
